requirements.txt
flask==3.0.0
replicate==0.24.0
Pillow==10.2.0
requests==2.31.0
